backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
networkx==3.2.1
scikit-learn==1.4.0
numpy==1.26.4
cohere==4.47
pydantic==2.6.1
python-dotenv==1.0.1
